Potential fix for the empty string test
Modified the if statement in `strip` that checks for empty strings. This
should ensure the test that checks if it works with empty strings will
now run properly, and the core will not be dumped.

`leak_report.md` will be updated shortly after this commit if this
change is successful.